fix: behavior of validationexception (part2)
